added map to enable copilot for 1 insert mode
I found the constant suggestions to be more distracting than useful, so
for now I'm keeping them disabled by default, and I've added this nice
mapping to enable copilot suggestions for the current insert mode until
you leave insert mode. Then the next insert will have no suggestions.

And if I want it back to how it was, I can still yoC to re-enable
Copilot to be on all the time.